Fixed the exceptions thrown, refactored boiler codes, fixed customer bug not saving if customer details is absent

+++ hustle-emarket-rest/src/main/java/org/emarket/hustle/hustleemarketrest/controller/StoreRestController.java
@@ -4,16 +4,25 @@ import java.util.List;
 
 import org.emarket.hustle.hustleemarketrest.BcryptSecurity;
 import org.emarket.hustle.hustleemarketrest.entity.Store;
+import org.emarket.hustle.hustleemarketrest.response.NotFoundException;
+import org.emarket.hustle.hustleemarketrest.response.ProcessConfirmation;
 import org.emarket.hustle.hustleemarketrest.service.StoreService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+/**
+ *
+ * there is no post/insert in Store
+ * because you need to be a Seller to create
+ * a Store
+ *
+ */
 
 @RestController
 @RequestMapping("/emarket-hustle")
@@ -26,8 +35,9 @@ public class StoreRestController
 	public BcryptSecurity bcrypt;
 
 	/*
-	 * BASIC CRUD OPERATIONS
-	 *
+	 * #######################################
+	 * ############ GET STORE ################
+	 * #######################################
 	 */
 
 	@GetMapping("/stores")
@@ -39,33 +49,37 @@ public class StoreRestController
 		}
 		catch (Exception e)
 		{
-			throw new RuntimeException("No stores were found");
-		}
-
-	}
-
-	@GetMapping("/stores/{id}")
-	public Store getSellerById(@PathVariable int id)
-	{
-		return storeService.getStoreById(id);
-	}
-
-	@PostMapping("/stores")
-	public Store addSeller(@RequestBody Store store)
-	{
-		store.setId(0);
-
-		try
-		{
-			storeService.saveStore(store);
-			return store;
-		}
-		catch (Exception e)
-		{
+			e.printStackTrace();
 			throw new RuntimeException(e);
 		}
 
 	}
+
+	/*
+	 * #######################################
+	 * ########### GET STORE BY ID ###########
+	 * #######################################
+	 */
+
+	@GetMapping("/stores/{id}")
+	public Store getSellerById(@PathVariable int id)
+	{
+		try
+		{
+			return storeService.getStoreById(id);
+		}
+		catch (Exception e)
+		{
+			e.printStackTrace();
+			throw new RuntimeException(e);
+		}
+	}
+
+	/*
+	 * #######################################
+	 * ########### UPDATE STORE ##############
+	 * #######################################
+	 */
 
 	@PutMapping("/stores")
 	public Store updateSeller(@RequestBody Store store)
@@ -77,19 +91,49 @@ public class StoreRestController
 		}
 		catch (Exception e)
 		{
+			e.printStackTrace();
 			throw new RuntimeException(e);
 		}
 
 		return store;
 	}
 
+	/*
+	 * #######################################
+	 * ######### DELETE STORE BY ID ##########
+	 * #######################################
+	 */
 	@DeleteMapping("/stores/{id}")
-	public String deleteStoreById(@PathVariable int id)
+	public ProcessConfirmation deleteStoreById(@PathVariable int id)
 	{
 
-		storeService.deleteStoreById(id);
+		try
+		{
+			Store store = storeService.getStoreById(id);
 
-		return ("Deleted Store with id: " + id);
+			if(store == null)
+			{
+				throw new NotFoundException("STORE WITH ID: " + id);
+			}
+
+			/*
+			 * severe the ties between the store and the seller
+			 * so we can safely delete the store
+			 */
+			store.getSeller().setSellerStore(null);
+
+			storeService.deleteStoreById(id);
+
+			return new ProcessConfirmation("SUCCESS", "SELLER",
+					"THE CUSTOMER WITH ID:" + id + " WAS DELETED.");
+		}
+		catch (Exception e)
+		{
+			e.printStackTrace();
+
+			throw new RuntimeException(e);
+		}
+
 	}
 
 }
